Docs + cleanup pass
