+++ 11ty/cheerio.ts
@@ -6,30 +6,30 @@ import { dirname, resolve } from "path";
 export { load } from "cheerio";
 
 /** Convenience function that combines readFile and load. */
-export async function loadFromFile(
+export const loadFromFile = async (
 	inputPath: string,
 	options?: CheerioOptions | null,
 	isDocument?: boolean
-) {
-	return load(await readFile(inputPath, "utf8"), options, isDocument);
-}
+) => load(await readFile(inputPath, "utf8"), options, isDocument);
 
 /**
  * Retrieves content for a data-include, either from _includes,
  * or relative to the input file.
+ * Operates synchronously for simplicity of use within Cheerio callbacks.
  *
  * @param includePath A data-include attribute value
  * @param inputPath Path (relative to repo root) to file containing the directive
  * @returns
  */
 function readInclude(includePath: string, inputPath: string) {
-	if (includePath.startsWith(".."))
-		return readFileSync(resolve(dirname(inputPath), includePath), "utf8");
+	const relativePath = resolve(dirname(inputPath), includePath);
+	if (includePath.startsWith("..")) return readFileSync(relativePath, "utf8");
 
 	try {
+		// Prioritize any match under _includes (e.g. over local toc.html built via XSLT)
 		return readFileSync(resolve("_includes", includePath), "utf8");
 	} catch (error) {
-		return readFileSync(resolve(dirname(inputPath), includePath), "utf8");
+		return readFileSync(relativePath, "utf8");
 	}
 }
 
@@ -46,7 +46,7 @@ function readInclude(includePath: string, inputPath: string) {
 export function flattenDom(content: string, inputPath: string) {
 	const $ = load(content);
 
-	$("body [data-include]").each((i, el) => {
+	$("body [data-include]").each((_, el) => {
 		const replacement = readInclude(el.attribs["data-include"], inputPath);
 		// Replace entire element or children, depending on data-include-replace
 		if (el.attribs["data-include-replace"]) $(el).replaceWith(replacement);
