+++ _includes/techniques/applicability.html
@@ -1,4 +1,4 @@
-{% # Autogenerated content appended to applicability section %}
+{% # Autogenerated content added to About box %}
 {% if techniqueAssociations %}
 	{% if techniqueAssociations.size == 1 %}
 	<p>
