+++ _includes/understanding/about.html
@@ -5,5 +5,5 @@
 {%- else if guideline.type == "Guideline" -%}
 	{% sectionbox "guideline" "Guideline" -%}
 		{{ guideline.content }}
-	{% endsectionbox %}
+	{%- endsectionbox %}
 {%- endif -%}
