+++ package.json
@@ -17,13 +17,13 @@
     "cheerio": "^1.0.0-rc.12",
     "glob": "^10.3.16",
     "liquidjs": "^10.13.1",
-    "lodash-es": "^4.17.21"
+    "lodash-es": "^4.17.21",
+    "tsx": "^4.11.0"
   },
   "devDependencies": {
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^20.12.12",
     "prettier": "^3.3.2",
-    "tsx": "^4.11.0",
     "typescript": "^5.4.5"
   }
 }
